#106 Do not use CentralDirectory when extract all
