+++ src/main/java/ru/olegcherednik/zip4jvm/io/in/file/random/BaseRandomAccessDataInput.java
@@ -16,9 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package ru.olegcherednik.zip4jvm.io.in;
+package ru.olegcherednik.zip4jvm.io.in.file.random;
 
 import ru.olegcherednik.zip4jvm.io.BaseMarker;
+import ru.olegcherednik.zip4jvm.io.in.BaseDataInput;
 
 import java.io.IOException;
