+++ src/main/java/ru/olegcherednik/zip4jvm/io/readers/zip64/Zip64Reader.java
@@ -21,7 +21,7 @@ package ru.olegcherednik.zip4jvm.io.readers.zip64;
 import ru.olegcherednik.zip4jvm.exception.SignatureNotFoundException;
 import ru.olegcherednik.zip4jvm.exception.Zip4jvmException;
 import ru.olegcherednik.zip4jvm.io.in.DataInput;
-import ru.olegcherednik.zip4jvm.io.in.RandomAccessDataInput;
+import ru.olegcherednik.zip4jvm.io.in.file.random.RandomAccessDataInput;
 import ru.olegcherednik.zip4jvm.model.Zip64;
 
 import java.io.IOException;
